move Hasura.Class to Hasura.Metadata.Class & document the module
