+++ server/src-lib/Hasura/Metadata/Class.hs
@@ -1,4 +1,5 @@
-module Hasura.Class
+-- | This module has type class and types which implements the Metadata Storage Abstraction
+module Hasura.Metadata.Class
   ( MetadataStorageT(..)
   , runMetadataStorageT
   , MonadMetadataStorage(..)
@@ -28,10 +29,46 @@ runMetadataStorageT
 runMetadataStorageT =
   runExceptT . unMetadataStorageT
 
+{- Note [Todo: Common interface for eventing sub-system]
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Scheduled and Table event triggers have their own implementation
+and tables in Postgres. The core logic for both is same. So it
+is necessary to have a unified interface (maybe via a Postgres extension)
+where all event triggers implementation can extend from. By this we can
+fairly reduce the Postgres schema foot print and interactions to the database.
+
+TODO: Reference to open issue or rfc?
+-}
+
+-- | Metadata storage abstraction via a type class.
+--
+-- This class has functions broadly related to:
+--
+-- 1. Metadata Management (TODO)
+-- ---------------------
+-- Basic metadata management functions such as retrieving metadata from storage
+-- database and replacing the given metadata.
+--
+-- 2. Scheduled Triggers
+-- ---------------------
+-- Eventing sub-system for scheduled triggers is implemented via metadata storage.
+-- All necessary functions are included in the type class. (TODO) The console has
+-- This also includes functions to fetch events and their invocations so that the
+-- console can show them in the UI.
+-- For more details, refer description in 'Hasura.Eventing.ScheduledTrigger' module.
+--
+-- 3. Async Actions (TODO)
+-- ----------------
+-- Operations to implement async actions sub-system. This includes recording an
+-- async action event and retreiving the details of action delivery to the webhook.
+-- For more details see Note [Async action architecture] in 'Hasura.GraphQL.Execute.Action' module.
 
 class (Monad m) => MonadMetadataStorage m where
 
   -- Scheduled triggers
+  -- By design, scheduled trigger eventing has many database interactions.
+  -- Hence we have many functions. We can reduce the functions by having
+  -- a common interface for eventing. See Note [Todo: Common interface for eventing sub-system]
   getDeprivedCronTriggerStats :: MetadataStorageT m [CronTriggerStats]
   getScheduledEventsForDelivery :: MetadataStorageT m ([CronEvent], [OneOffScheduledEvent])
   insertScheduledEvent :: ScheduledEventSeed -> MetadataStorageT m ()
